Fix requirement on argc
This was initially done in f0eb507ee142f45b08533a57744c516aa529d5b4 by
@polvalente but introduced an unforseen bug.

Building on their example, this again removes the requirement for argc
for the desired behavior of being able to run changes locally for
testing.

Also adds a bit of testing for the args and results.