Move launcher code out of priv
Even though we're telling people to set `runtime: false` when they
depend on `:bakeware`, it's easy to forget. If you forget, then the
launcher code is included twice - once in the right place and again
inside of the OTP release. By moving the launcher code outside of the
`priv` directory, it no longer gets included in the release for
`runtime: true`.

This also lets us experiment with including runtime code in Bakeware
without bloating the generated binaries as much. For reference, here are
the current release contents for `runtime: true`:

```
bakeware-0.1.1
└── ebin
    ├── bakeware.app
    ├── Elixir.Bakeware.Assembler.beam
    ├── Elixir.Bakeware.beam
    ├── Elixir.Bakeware.Script.beam
    └── Elixir.Mix.Tasks.Bakeware.Assemble.beam
```