Add test for simultaneous extraction
This tests the case where a user runs the extractor multiple times
simultaneously when nothing is in the cache. It would be bad if the
extractors stomped on each other. The design where extraction happens in
a unique temp directory and then does an atomic move to the run location
should prevent this. It's good to verify, though.