Use initial 'for' loop delcarations
Fixes #122

Compilation might fail if a system is not using C99 mode. So this just
changes these conditions to conform to that standard.

Note that this might also be satisfied by `CFLAGS+=-std=gnu99 mix compile`
or adding that flag in the Makefile, but its easy enough to change the few places here.